Font additions, Terminal UI Redesign, Integration for Window
Created basic terminal UI, the input field may be redesigned, if we want text, or buttons, etc. Moreover, we imported and generated a font asset for TMPro with Roboto Mono, a monospace font for the terminal with an apache license. Included in font folder is the license which we can add to credits for proper and legal use.